commit 2

+++ ASSIGNMENT 5/source/Systick_timer.h
@@ -8,10 +8,15 @@
 #ifndef SYSTICK_TIMER_H_
 #define SYSTICK_TIMER_H_
 
+/* Header files*/
+
 #include "logger.h"
 
+
+/* Function prototype*/
+
 void Init_SysTick(void);
-uint64_t current_time();
+uint32_t current_time();
 
 
 #endif /* SYSTICK_TIMER_H_ */
